reorganize tests a bit; add exclusions
