+++ nightwatch.conf.cjs
@@ -15,9 +15,11 @@ module.exports = {
   // An array of folders (excluding subfolders) where your tests are located;
   // if this is not specified, the test source must be passed as the second argument to the test runner.
   src_folders: [
-    'test',
-    // 'nightwatch/examples'
+    'test/e2e',
   ],
+
+  // Specify exclusions
+  exclude: ["test/e2e/lib"],
 
   // See https://nightwatchjs.org/guide/concepts/page-object-model.html
   page_objects_path: [],
