+++ test/e2e/lib/index.cjs
