+++ vite.config.ts
@@ -24,6 +24,14 @@ export default defineConfig(({ command, mode, ssrBuild }) => {
     // Test configuration for vitest
     test: {
       environment: 'jsdom', 
+      exclude: [
+        '**/node_modules/**', 
+        '**/dist/**', 
+        '**/cypress/**', 
+        '**/.{idea,git,cache,output,temp}/**', 
+        '**/{karma,rollup,webpack,vite,vitest,jest,ava,babel,nyc,cypress,tsup,build}.config.*',
+        '**/test/e2e/**'
+      ],
     },
   }
 })
